sqlalchemy dialect refactor
